refactor: 인증 상세 페이지 이의 신청 입력 팝업창 컴포넌트로 변경
